feat: add table a.k.a hashmap

+++ src/long_value.h
@@ -8,7 +8,7 @@
 
 typedef struct
 {
-    uint8_t capacity;
+    uint32_t capacity;
     uint32_t count;
 
     Value *values;
